+++ CHANGELOG.md
@@ -97,6 +97,7 @@ With this release the systemd configuration files for InfluxDB will use the syst
 - [#6819](https://github.com/influxdata/influxdb/issues/6819): Database unresponsive after DROP MEASUREMENT
 - [#6796](https://github.com/influxdata/influxdb/issues/6796): Out of Memory Error when Dropping Measurement
 - [#6946](https://github.com/influxdata/influxdb/issues/6946): Duplicate data for the same timestamp
+- [#7043](https://github.com/influxdata/influxdb/pull/7043): Remove limiter from walkShards
 
 ## v0.13.0 [2016-05-12]